requirements.txt
python-telegram-bot==20.8
apscheduler==3.10.1
pytz
requests
fastapi
uvicorn
